Modifcations for multiple databases and new install
git-svn-id: https://everydevel.svn.sourceforge.net/svnroot/everydevel/trunk@950 b9a9f069-900d-0410-9e71-c61feadb8a10

+++ ecore/tables/Pg/container.sql
@@ -7,7 +7,7 @@
 --
 
 CREATE TABLE "container" (
-  "container_id" serial NOT NULL,
+  "container_id" bigint NOT NULL,
   "context" text,
   "parent_container" bigint,
   PRIMARY KEY ("container_id")
